fix warning
[r26301]

+++ ir/ir/irprog.c
@@ -333,7 +333,7 @@ void add_irp_opcode(ir_op *opcode) {
 
 /* Removes opcode from the list of opcodes and shrinks the list by one. */
 void remove_irp_opcode(ir_op *opcode) {
-	assert(opcode->code < ARR_LEN(irp->opcodes));
+	assert((int) opcode->code < ARR_LEN(irp->opcodes));
 	irp->opcodes[opcode->code] = NULL;
 }
